增加了横电感和状态机
但是跑得更烂了XD

+++ src/drivers/debug.h
@@ -1,4 +1,4 @@
-extern unsigned int OutData[4];  
+extern unsigned short OutData[4];  
 unsigned short CRC_CHECK(unsigned char *Buf, unsigned char CRC_CNT);
 void OutPut_Data(unsigned short int*);
 void GetData(float Ch1,float Ch2,float Ch3,float Ch4,unsigned short int*);
